requirements.txt
Django==4.2.16
asgiref==3.8.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.5
pytz==2023.3.post1
sqlparse==0.5.1
flake8==5.0.4
tzdata==2023.4
psycopg==3.1.12
psycopg-binary==3.1.12
drf-spectacular==0.27.2
djangorestframework-simplejwt
pillow==11.0.0
